pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "psp"
version = "0.4.0"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1",
    "requests>=2.31",
    "rich>=13.0",
    "pyyaml>=6.0",
]

[project.scripts]
psp = "psp.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
